UPdate scripts and readme for issuing assets
