+++ scripts/setup-issuer.js
